file upload to /uploads/comics
made route folder a more tidy - moved multer code to its own config file

files now upload to correct directory
